// assets/src/pages/user/manage/data.d.ts
export interface Member {
  avatar: string;
  name: string;
  id: string;
}

export interface BasicListItemDataType {
  oaid: number;
  uid: number;
  username: string;
  nickname: string;
  token: string;
  secret: string;
  email: string;
  avatarUrl: string;
  webUrl: string;
  state: string;
  hash: string;
  create_time: number;
  update_time: number;
  access: string;
  password: string;
}